change(models): nit y rut a INT

+++ src/main/java/com/itm/ecosurprise/models/Comerciante.java
@@ -19,8 +19,8 @@ import jakarta.persistence.Table;
 @Table(name = "comerciantes")
 public class Comerciante extends Usuario {
 
-	private long NIT;
-	private long RUT;
+	private int NIT;
+	private int RUT;
 	
 	@OneToMany
 	private List<Sede> sedes;
